Add docker files

+++ backend/src/main/java/com/pku/service/UserService.java
@@ -6,5 +6,5 @@ import org.springframework.http.HttpStatus;
 
 public interface UserService {
     public HttpStatus registerUser(UserDTO userDTO);
-    public Integer loginUser(UserDTO userDTO);
+    public Long loginUser(UserDTO userDTO);
 }
